Re-introduce original media file and start treating as a back-compatibility layer
